# arg_auth/src/api-schema/generated/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/callback": {
      "get": {
        "description": "Frontend usage: api.auth.callback()",
        "operationId": "callback",
        "parameters": [
          {
            "name": "code",
            "required": false,
            "in": "query",
            "description": "Microsoft Code",
            "schema": {
              "example": "abc",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of Reports Menu",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "message": {
                        "type": "Redirect",
                        "example": "Redirect"
                      }
                    }
                  },
                  "required": [
                    "items"
                  ]
                }
              }
            }
          }
        },
        "summary": "Set Cookie and Redirect (Method: callback)",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth": {
      "post": {
        "description": "Frontend usage: api.auth.validateUser()",
        "operationId": "validateUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "authData"
                ],
                "properties": {
                  "authData": {
                    "type": "object",
                    "required": [
                      "route"
                    ],
                    "properties": {
                      "route": {
                        "type": "string",
                        "example": "N",
                        "required": false
                      },
                      "method": {
                        "type": "string",
                        "example": "N",
                        "required": false
                      },
                      "restApi": {
                        "type": "boolean",
                        "example": true,
                        "required": true
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User Details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "email": {
                        "type": "string",
                        "example": "test@amref.com"
                      },
                      "extensionAttribute1": {
                        "type": "string",
                        "example": "TEST"
                      }
                    }
                  },
                  "required": [
                    "items"
                  ]
                }
              }
            }
          }
        },
        "summary": "Authenticate and Authorize User (Method: validateUser)",
        "tags": [
          "auth"
        ]
      }
    }
  },
  "info": {
    "title": "ARG API",
    "description": "The ARG API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}